+++ OUDS/Core/Components/Sources/Dialogs/Alert/Internal/AlertMessageBackgroundModifier.swift
@@ -1,0 +1,49 @@
+//
+// Software Name: OUDS iOS
+// SPDX-FileCopyrightText: Copyright (c) Orange SA
+// SPDX-License-Identifier: MIT
+//
+// This software is distributed under the MIT license,
+// the text of which is available at https://opensource.org/license/MIT/
+// or see the "LICENSE" file for more details.
+//
+// Authors: See CONTRIBUTORS.txt
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
+//
+
+import OUDSTokensSemantic
+import SwiftUI
+
+struct AlertMessageBackgroundModifier: ViewModifier {
+
+    // MARK: - Properties
+
+    let status: OUDSAlertMessage.Status
+
+    @Environment(\.theme) private var theme
+
+    // MARK: - Body
+
+    func body(content: Content) -> some View {
+        content.oudsBackground(color)
+    }
+
+    // MARK: - Helpers
+
+    private var color: MultipleColorSemanticToken {
+        switch status {
+        case .neutral:
+            theme.colors.surfaceSecondary
+        case .accent:
+            theme.colors.surfaceStatusAccentMuted
+        case .positive:
+            theme.colors.surfaceStatusPositiveMuted
+        case .warning:
+            theme.colors.surfaceStatusWarningMuted
+        case .negative:
+            theme.colors.surfaceStatusNegativeMuted
+        case .info:
+            theme.colors.surfaceStatusInfoMuted
+        }
+    }
+}