+++ OUDS/Core/ThemesContract/Sources/_ThemesContract.docc/Tutorial/swift-samples/tutorial-5-9.swift
@@ -17,7 +17,7 @@ import SwiftUI
 @main
 struct OUDSTutorialSandboxApp: App {
 
-    let myTheme = OrangeTheme(fontFamily: "Menlo", tuning: Tuning(hasRoundedButtons: true, hasRoundedTextInputs: true))
+    let myTheme = OrangeTheme(fontFamily: "Menlo", tuning: Tuning(hasRoundedButtons: true, hasRoundedTextInputs: true, hasRoundedAlertMessages: true))
 
     var body: some Scene {
         WindowGroup {